Factor out requirement graph builder code into its own file
